Make locks more friendly (#1539)
### Issue
#973 

### Description
User currently doesn't know that they're waiting on a lock to be
released. This waiting on the lock is perceived as a hang. This change
does a small timed wait of 1sec, and if lock wasn't obtained the log
will print out that the thread is blocked on the lock and also which PID
currently owns it.

### List of the changes
- Changed mutex wrapped structure to contain owner PID
- Changed the way we lock the mutex, so that the first attempt fails
after 1sec and we get a chance to log something.
- The initialisation code doesn't need to change, since the code already
tries to resize the underlying file. So this should happen automatically
on all systems.

### Testing
I've manually ran MultipleThreadsMultipleClustersOpenClose and saw the
logs.

### API Changes
There are no API changes in this PR.