Use correct variable for clang-tidy path (#1611)
### Issue
/

### Description
This pull request makes a small update to the `CMakeLists.txt` file,
ensuring that the correct `.clang-tidy-build` configuration file is used
during static analysis. This allows the project to be consumed safely as
a submodule.

### List of the changes
- Updated the `CMAKE_CXX_CLANG_TIDY` and `CMAKE_C_CLANG_TIDY` commands
to use `${CMAKE_CURRENT_SOURCE_DIR}` instead of `${PROJECT_SOURCE_DIR}`
for the `--config-file` argument, ensuring the correct
`.clang-tidy-build` file is referenced in each source directory.

### Testing
Manual

### API Changes
/