Make clang-tidy IDE in sync with clang-tidy build (#1616)
### Issue
/

### Description
This pull request updates the project's static analysis configuration
for clang tools, focusing on improving naming conventions and ensuring
that the correct configuration files are used for Clang-Tidy during
builds.

### List of the changes
* Updated `.clangd` to only enable the `readability-identifier-naming`
check (removed the global disable-all `-*` which doesn't work in
clangd), and changed function, parameter, and constant parameter naming
rules from `snake_case` to `lower_case` because this is the actual rule.
* Renamed .clang-tidy-build to .clang-tidy

### Testing
Manual

### API Changes
/