Skip making a copy of the ttsim binary for single chip workloads (#1648)
### Issue
(Link to Github issue(s))

### Description
Copying the simulator binary is only required for multi-chip workloads
in the same process, so we are creating unnecessary copies in the single
chip case.

### List of the changes
Skip copying the sim binary when the cluster is only a single chip.

### Testing
Local testing.

### API Changes
(When making API changes, don't merge this PR until tt_metal and
tt_debuda PRs are approved.)
(Then merge this PR, change the client PRs to point to UMD main, and
then merge them.)
(Remove this line if untrue) There are no API changes in this PR.
(Remove following lines if untrue) This PR has API changes:
- [ ] (If breaking change) tt_metal approved PR pointing to this branch:
link
- [ ] (If breaking change) tt_debuda approved PR pointing to this
branch: link