Fix Tsan/Asan python builds (#1652)
### Issue
On a path to https://github.com/tenstorrent/tt-umd/issues/636

### Description
Fix python tests in case of asan and tsan, which currently don't work if
you manually run
https://github.com/tenstorrent/tt-umd/actions/workflows/build-and-run-all-tests.yml

### List of the changes
- Change the order of tests, so that we first install the wheel and run
python tests, and then download build artefacts and run cpp unit tests.
This is because the libdevice.so might get loaded instead of the one
from the wheel. Hopefully the #1340 might help in general with this
issue
- Block building python lib in case of asan/tsan
- Don't build python lib when building unit tests, they are unnecessary
- Override Asan and Tsan to Release when building python wheel in our CI

### Testing
CI TSAN tests, with some other changes combined:
https://github.com/tenstorrent/tt-umd/actions/runs/19664175096

### API Changes
There are no API changes in this PR.