Change EXPECT to ASSERT in sysmem tests (#1670)
### Issue
Uncovered by failure in
https://github.com/tenstorrent/tt-umd/actions/runs/19768960443/job/56649021362?pr=1669

### Description
When sysmem is setup bad, likely the whole full 1 megabyte will be bad.
Printing it for every byte just makes a ton of longs on CI without any
benefit. Changing this to ASSERT will stop the test after the first
failure

### List of the changes
- Change EXPECT_EQ to ASSERT_EQ in each for loop which goes over 1
megabyte.

### Testing
No testing

### API Changes
There are no API changes in this PR.