Use 1MB for WH cached TLB (#1669)
### Issue
Part of https://github.com/tenstorrent/tt-umd/issues/1252
Issue with tt-exalens tests creating many drivers.

### Description
It seems like we have a hard limit on live drivers which is controlled
by the number of TLBs.
WH has only 10 2MB tlbs, but 150 of 1MB tlbs, so this will greatly
increase the limit.
In any case, we shouldn't have that many drivers live.

### List of the changes
- Protected get_cached_tlb_size function, overridden to 1MB for WH and
2MB for BH
- While at it, I also added get_static_vc function, which was the only
other place there was arch based code in tt_device
- @joelsmithTT found an issue with sysmem buffer, we weren't properly
addressing the right tlb, rather the tlb 0, so now that we started using
the tlb 0 through tt_device this was clashing

### Testing
tt-exalens tests passed with this change

### API Changes
There are no API changes in this PR.