Add standalone ASIO library (#1674)
### Issue
#1351 

### Description
This pull request introduces the standalone version of the Asio library
into the project and updates the build and test configuration to support
it. The main changes involve adding Asio as a dependency, configuring
its usage in the build system, and including it in relevant test files.

### List of the changes
**Dependency management and build configuration:**

* Added Asio as a standalone dependency in `third_party/CMakeLists.txt`
using `CPMAddPackage`, specifying version 1.30.2 and disabling Boost and
deprecated features. Also set `ASIO_INCLUDE_DIR` for use in the build.
* Updated `device/CMakeLists.txt` to include the Asio header path via
`target_include_directories` and defined `ASIO_STANDALONE` for the
`device` target to ensure correct compilation without Boost.

### Testing
Manual

### API Changes
/