Add TopologyDiscovery discover method (#1654)
### Issue
Related to https://github.com/tenstorrent/tt-umd/issues/1548

### Description
Changes are in progress for TopologyDiscovery method to return a set of
TTDevices.
This also enables easier code not just for exalens, but tooling as well.
This is general direction where cpp code is also going in, but this
change makes it easier to earlier integrate python clients, the api will
remain the same.

### List of the changes
- Add TopologyDiscovery discover method which returns both
ClusterDescriptor and set of initialized TTDevices
- Simplify some python tests
- Fixed no_remote_discovery usage in TopoDiscovery

### Testing
Tested this as part of https://github.com/tenstorrent/tt-smi/pull/142

### API Changes
This PR has API changes, adds another option for python api

---------

Co-authored-by: Aleksa Markovic <aleksamarkovic@tenstorrent.com>