Verify ERISC FW hash in TopologyDiscovery (#1613)
### Issue
#1612 #1637

### Description
UMD will now calculate a SHA256 hash of the read only segment of ETH FW
containing the .text section.
The range for WH is verified to be `0x2000-0x8FFF` by the WH ERISC FW
linker script.
The range for BH starts at `0x70000` and the end depends on the size of
the text segment of the ERISC FW.
A warning is logged if the hash from the read L1 of the core does not
match the pre-calculated hash and the core will not be used for remote
chip discovery.

### List of the changes
- Introduce third party dependency picosha2
- Added hashes of most supported WH and BH ERISC FW versions
- Verify ETH FW hash for WH and BH as a part of TopologyDiscovery

### Testing
CI

### API Changes
There are no API changes in this PR.