Unify SPDX headers (#1645)
### Issue


### Description
I realized that our SPDX headers are half // half *, so this change
unifies them.
Also add an empty line after it.
This matches how they look in tt-metal as well

### List of the changes
- Change the check so that default license comments are with //
- Change existing comments from /**/ to //
- Add lines to existing comments in cpp and python
- Add (c) to python spdx header
- Change all (c) with copyright symbols

### Testing
Tested manually that a new file would get the same header automatically.

### API Changes
There are no API changes in this PR.