Implement SysmemManager for simulation (#1667)
### Issue

/

### Description

Implement SysmemManager for simulation chips. The simulation doesn't
have the actual system memory, we can't actually allocate it, so we need
to simulate it by simply creating a vector of the size that is equal to
number of host memory channels. Functions in Chip class need to
overridden in order to fully support integration of SysmemManager inside
SimulationChip.

### List of the changes

- Implement SimulationSysmemManager class
- Integrate it inside SimulationChip
- Override functions in SimulationChip that are needed to support system
memory in simulation
- Add tests for SimulationSysmemManager

### Testing
Added new test. E2E tests to be run by Matt

### API Changes
/

---------

Co-authored-by: Bojan Roško <156314064+broskoTT@users.noreply.github.com>