Fail test if package is not found (#1738)
### Issue
/

### Description
This pull request updates the CI workflow to ensure that missing Python
packages cause a build failure rather than being silently skipped. This
change helps catch issues earlier in the development process.

### List of the changes
CI workflow enforcement:

*
[`.github/workflows/build-device.yml`](diffhunk://#diff-51443dce697ed156c7d3fa97c7e64ba19f632eca72c5c178ea640bf6961c015dL395-R396):
Changed the behavior when the Python package is not found—now the script
prints an error message and exits with a failure code to ensure CI fails
if the package is missing.


### Testing
/

### API Changes
/